# backend/openapi.yaml
openapi: 3.1.0
info:
  title: 'Ragab CRM API'
  version: 1.0.0
  description: 'Ragab CRM API. Errors are RFC 9457 application/problem+json documents whose `code` member is a stable `module.condition` identifier. Write endpoints require an Idempotency-Key header (ULID or UUID).'
servers:
  -
    url: /api/v1
    description: 'Relative to the deployment host, e.g. http://localhost:8000 locally.'
paths:
...
  /auth/session:
    get:
      operationId: auth.session
      summary: "Session timing, so the frontend can warn before a session lapses rather\nthan discovering it on the next request"
      tags:
        - Auth
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  inactivity_minutes:
                    type: integer
                  authenticated:
                    type: boolean
                required:
                  - inactivity_minutes
                  - authenticated
        default:
          description: 'An RFC 9457 problem document.'
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
...
  '/departments/{department}/deactivate':
    post:
      operationId: departments.deactivate
      summary: 'Deactivate — refused while active tickets remain'
      tags:
        - Departments
      parameters:
        -
          name: department
          in: path
          required: true
          description: 'The department ID'
          schema:
            type: integer
        -
          name: Idempotency-Key
          in: header
          required: true
          description: 'A ULID or UUID that identifies this write attempt. Repeating a request with the same key replays the stored response instead of acting twice; reusing a key with a different body returns 409. Keys are retained for 24 hours.'
          schema:
            type: string
            maxLength: 40
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  name:
                    type: string
                  is_active:
                    type: boolean
                required:
                  - id
                  - name
                  - is_active
        '404':
          $ref: '#/components/responses/ModelNotFoundException'
        '401':
          $ref: '#/components/responses/AuthenticationException'
        default:
          description: 'An RFC 9457 problem document.'
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '409':
          description: 'The Idempotency-Key was already used for a different request (code: platform.idempotency_conflict).'
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '425':
          description: 'A concurrent request with the same Idempotency-Key is still in flight (code: platform.idempotency_in_flight).'
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
...
  /profile/password:
    post:
      operationId: profile.password
      summary: 'Change password'
      tags:
        - Profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                required:
                  - status
        '401':
          $ref: '#/components/responses/AuthenticationException'
        '422':
          $ref: '#/components/responses/ValidationException'
        '403':
          $ref: '#/components/responses/AuthorizationException'
        default:
          description: 'An RFC 9457 problem document.'
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
...
  '/users/{user}/deactivate':
    post:
      operationId: users.deactivate
      summary: 'Deactivate, never delete'
      tags:
        - Users
      parameters:
        -
          name: user
          in: path
          required: true
          description: 'The user ID'
          schema:
            type: integer
        -
          name: Idempotency-Key
          in: header
          required: true
          description: 'A ULID or UUID that identifies this write attempt. Repeating a request with the same key replays the stored response instead of acting twice; reusing a key with a different body returns 409. Keys are retained for 24 hours.'
          schema:
            type: string
            maxLength: 40
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  name:
                    type: string
                  email:
                    type: string
                  role:
                    type:
                      - string
                      - 'null'
                  department_id:
                    type:
                      - integer
                      - 'null'
                  is_active:
                    type: boolean
                required:
                  - id
                  - name
                  - email
                  - role
                  - department_id
                  - is_active
        '404':
          $ref: '#/components/responses/ModelNotFoundException'
        '401':
          $ref: '#/components/responses/AuthenticationException'
        default:
          description: 'An RFC 9457 problem document.'
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '409':
          description: 'The Idempotency-Key was already used for a different request (code: platform.idempotency_conflict).'
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '425':
          description: 'A concurrent request with the same Idempotency-Key is still in flight (code: platform.idempotency_in_flight).'
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
components:
  schemas:
    ChangePasswordRequest:
      type: object
      properties:
        current_password:
          type: string
          description: "Proves the person at the keyboard is the account holder, not\nsomeone who found an unlocked screen."
        password:
          type: string
        password_confirmation:
          type: string
      required:
        - current_password
        - password
        - password_confirmation
      title: ChangePasswordRequest
...
    Problem:
      type: object
      title: Problem
      description: 'RFC 9457 problem details. Every 4xx and 5xx response in this API has this shape.'
      properties:
        type:
          type: string
          format: uri
          description: 'A URI identifying the problem type. Always the base URI plus `code`.'
        title:
          type: string
          description: 'A short, human-readable summary.'
        status:
          type: integer
          minimum: 400
          maximum: 599
        detail:
          type:
            - string
            - 'null'
          description: 'A human-readable explanation specific to this occurrence.'
        instance:
          type: string
          description: 'The request URI this problem occurred on.'
        code:
          type: string
          pattern: '^[a-z0-9_]+\.[a-z0-9_]+$'
          description: 'Stable machine identifier shaped `module.condition`. Branch on this, never on `title`.'
          examples:
            - platform.internal_error
            - platform.validation_failed
            - platform.unauthorized
            - platform.forbidden
            - platform.not_found
            - platform.method_not_allowed
            - platform.conflict
            - platform.too_many_requests
            - platform.request_failed
            - platform.idempotency_conflict
            - platform.idempotency_in_flight
        trace_id:
          type: string
          description: 'Correlation id for this request; matches the X-Request-Id response header and the request_id in the logs.'
        errors:
          type: object
          description: 'Present on validation failures: field name to list of messages.'
          additionalProperties:
            type: array
            items:
              type: string
      required:
        - type
        - title
        - status
        - instance
        - code
        - trace_id
  responses:
    ValidationException:
      description: 'Validation error'
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                description: 'Errors overview.'
              errors:
                type: object
                description: 'A detailed description of each field that failed validation.'
                additionalProperties:
                  type: array
                  items:
                    type: string
            required:
              - message
              - errors
    AuthenticationException:
      description: Unauthenticated
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                description: 'Error overview.'
            required:
              - message
    AuthorizationException:
      description: 'Authorization error'
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                description: 'Error overview.'
            required:
              - message
    ModelNotFoundException:
      description: 'Not found'
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                description: 'Error overview.'
            required:
              - message